refactor(backend-users): remove admin method
